starter/requirements.txt
numpy
pandas
scikit-learn
pytest
requests
fastapi==0.63.0
uvicorn
gunicorn
matplotlib
seaborne
